add progress bar to complaint uploads
